Update MCTS to use prior value function without rollout

+++ Caro_pybind/Caro_pybind/MCTS_pybind.cpp
@@ -24,7 +24,7 @@ PYBIND11_MODULE(MCTS_pybind, m)
 		.def("__str__", &TreeNode::to_string);
 
 	py::class_<MCTS_AI>(m, "MCTS_AI")
-		.def(py::init<int, int, int, Caro const&, std::string, int, std::function<double(std::array<std::array<int, 30>, 30>, int)>, int>(),
+		.def(py::init<int, int, int, Caro const&, std::string, int, std::function<float(std::array<std::array<int, 30>, 30>, int)>, int>(),
 			py::arg("_player"), py::arg("_min_visits"), py::arg("_n_sim"), py::arg("_board"), 
 			py::arg("_mode") = "greedy", py::arg("_ai_moves_range") = 1, py::arg("_eval") = nullptr, py::arg("_prior_strength") = -1)
 		.def("get_move", &MCTS_AI::get_move)
